+++ packages/components/forms/src/ValidationMessage/ValidationMessage.tsx
@@ -13,8 +13,7 @@ import { useFormControl } from '../FormControl/FormControlContext';
 import { useDensity } from '@contentful/f36-utils';
 
 export interface ValidationMessageInternalProps
-  extends CommonProps,
-    MarginProps {
+  extends CommonProps, MarginProps {
   children: React.ReactNode;
 }