+++ packages/components/button/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 6.7.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-core@6.7.2
+  - @contentful/f36-spinner@6.7.2
+  - @contentful/f36-tooltip@6.7.2
+
 ## 6.7.1
 
 ### Patch Changes