+++ packages/components/image/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 6.7.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-core@6.7.2
+  - @contentful/f36-skeleton@6.7.2
+
 ## 6.7.1
 
 ### Patch Changes